# docs/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  internal_handler.createSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  subscription-service_internal_model.Subscription:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      end_date:
        type: string
      id:
        type: integer
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      updatedAt:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API сервис для управления онлайн-подписками
  title: Subscription Service API
  version: "1.0"
paths:
  /subscriptions:
    get:
      description: Возвращает список всех подписок
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/subscription-service_internal_model.Subscription'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить все подписки
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Создание новой подписки для пользователя
      parameters:
      - description: Данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/internal_handler.createSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/subscription-service_internal_model.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать подписку
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Удаление подписки по ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить подписку
      tags:
      - subscriptions
    get:
      description: Возвращает подписку по её идентификатору
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscription-service_internal_model.Subscription'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить подписку по ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Обновление данных существующей подписки
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      - description: Обновлённые данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/internal_handler.createSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscription-service_internal_model.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить подписку
      tags:
      - subscriptions
  /subscriptions/sum:
    get:
      description: Возвращает суммарную стоимость подписок за период с фильтрацией
        по user_id и service_name
      parameters:
      - description: UUID пользователя
        in: query
        name: user_id
        type: string
      - description: Название сервиса
        in: query
        name: service_name
        type: string
      - description: Дата начала периода (MM-YYYY)
        in: query
        name: start_date
        required: true
        type: string
      - description: Дата окончания периода (MM-YYYY)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: sum
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Подсчитать сумму подписок
      tags:
      - subscriptions
swagger: "2.0"